Remove DafnyPipline -> DafnyRuntime dependency
This is a false dependency, only technically required because of an unused “using” declaration (now removed). It’s also confusing because it’s the compiled code that depends on the C# runtime, and only in the C# case. We want to control that meta-dependency in a more explicit and manual way to ensure compiled code compatibility.

The individual target compilers could have a build-time dependency on the runtimes, in order to refrerence their contents and ensure the compiled code uses correct symbols, but that will still be a different flavour of dependency.